[FileTableDialog] add note related to updating entries from FileTableDB
Specifically, if a project has changed the meaning of entries (e.g. using file IDs that are leftover from another game), we shouldn't be changing the filetype to the original.

The only decent way I can think of handling this is via a dialog that comes up for each different/colliding instance. (Probably ignore any Binary -> * conversions though??)